+++ template.h
@@ -9,6 +9,7 @@
   </head>\
   <body>"
 
+#define DOWNLOAD_HTML_TAIL_TEMPLATE "</body>\n</html>"
 #define UPLOAD_HTML_TEMPLATE \
   "<!DOCTYPE html>\
 <html>\
@@ -20,7 +21,7 @@
   </head>\
   <body>\
     <form> <input type =\"file\" id=\"avatar\" name=\"avatar\">\
-    <input type=\"text\" id=\"path\" name=\"path\">\
+    <input type=\"text\" id=\"path\" name=\"path\" placeholder=\"上传路径，默认为/\">\
     <button id=\"btn\" type=\"button\">Upload</button>\
     </form>\
   </body>\
@@ -28,6 +29,9 @@
     $('button').click(function(){\
       var files = $('#avatar').prop('files');\
       var path = $('#path').val();\
+      if(path === '') {\
+        path = '/';\
+      }\
       var data = new FormData();\
       data.append('avatar', files[0]);\
       data.append('path', path);\
